Refactored out code to read and filter household power data into a function.  Added a Makefile to download the data from url and remake figures as needed.
This version is minimally useful.  If this needed to go further, we would work on making the filtering code do filtering out
of memory, so could work with truly large data.  Also restructure so only have to read in 1 time, though this might not
fit the project requirements which wanted each plotX.R file to read in the data.